// src/app/court/models/CourtAffidavitUserResponse.ts
export class CourtAffidavitUserResponse {
  userAffidavitId: number;
  customName: string;
  status: string;
  temName: string;
  price: number;
  userId: number;
  userName: string;
  address: string;
  email: string;
  contact: string;
  registrarStatus: string;
  registrarComments: string;
  registrarName:string;
  registrarCourtName:string;
  registrarId:number;
  missedAppointment: boolean;
  isUploaded:Boolean;
  isCaseRelated:Boolean;
}